Further structural changes to handler
