+++ src/handler/node_info.rs
@@ -1,0 +1,50 @@
+/// This type relaxes the requirement of having an ENR to connect to a node, to allow for unsigned
+/// connection types, such as multiaddrs.
+pub enum NodeContact {
+    /// We know the ENR of the node we are contacting.
+    Enr(Enr)
+    /// We don't have an ENR, but have enough information to start a handshake. 
+    ///
+    /// The handshake will request the ENR at the first opportunity.
+    /// The public key can be derived from multiaddr's whose keys can be inlined. The `TryFrom`
+    /// implementation for `String` and `MultiAddr`.
+    Raw {
+        /// An ENR compatible public key, required for handshaking with peers.
+        public_key: CombinedPublicKey,
+        /// The socket address and ModeId of the peer to connect to.
+        NodeAddress(NodeAddress)
+    }
+}
+
+impl NodeContact {
+    pub fn is_enr(&self) -> bool {
+        match self {
+            Enr(_) => true
+                _ => false
+        }
+    }
+
+    pub fn udp_socket(&self) -> Result<SocketAddr, &'static str> {
+        match self {
+            Enr(ref enr) => enr.udp_socket().map_err(|_| "ENR does not contain an IP and UDP port")?
+            NodeAddress(ref node_address) => node_address.socket_addr.clone()
+        }
+    }
+}
+
+/// A representation of an unsigned contactable node.
+pub struct NodeAddress {
+    /// The destination socket address.
+    socket_addr: SocketAddr,
+    /// The destination Node Id. 
+    node_id: NodeId
+}
+
+impl NodeAddress {
+    pub fn new(socket_addr: SocketAddr, node_id: NodeId) -> Self {
+        Self {
+            socket_addr,
+            node_id
+        }
+    }
+}
